requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
asgiref==3.8.1
astunparse==1.6.3
attrs==23.2.0
awscli==1.33.26
boto3==1.34.144
botocore==1.34.144
certifi==2024.6.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
coverage==7.6.0
cryptography==44.0.0
cycler==0.12.1
distlib==0.3.9
Django==5.1.5
dnspython==2.6.1
docker==7.1.0
docutils==0.16
email_validator==2.2.0
fastapi==0.89.1
fastapi-cli==0.0.4
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.0
fsspec==2024.6.1
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.64.1
gunicorn==20.1.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
hydra-core==1.3.2
idna==3.7
imageio==2.34.2
iniconfig==2.0.0
intel-openmp==2021.4.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.4
jmespath==1.0.1
keras==3.4.1
keyboard==0.13.5
keyring==8.7
keyrings.alt==5.0.1
kiwisolver==1.4.5
libclang==18.1.1
loguru==0.7.2
lxml==5.3.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mediapipe==0.10.14
mkl==2021.4.0
ml-dtypes==0.3.2
more-itertools==10.3.0
mpmath==1.3.0
mss==9.0.1
namex==0.0.8
networkx==3.3
Nuitka==2.3.11
numpy==1.26.4
omegaconf==2.3.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
optree==0.11.0
ordered-set==4.1.0
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pdfminer==20191125
pdfminer.six==20240706
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
protobuf==4.25.3
psutil==6.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==1.10.17
pygame==2.6.0
Pygments==2.18.0
pyparsing==3.1.2
pypdf==5.2.0
PyPDF2==3.0.1
pytest==8.3.1
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
reportlab==4.2.5
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
roboflow==1.1.34
rsa==4.7.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.2
scipy==1.14.0
seaborn==0.13.2
sentry-sdk==2.10.0
setuptools==70.2.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.4.7
sqlparse==0.5.3
starlette==0.22.0
sympy==1.12.1
tabulate==0.9.0
tbb==2021.13.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-intel==2.16.2
termcolor==2.4.0
thop==0.1.1.post2209072238
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
ultralytics==8.2.64
ultralytics-thop==2.0.0
urllib3==2.2.2
uvicorn==0.20.0
virtualenv==20.29.1
voluptuous==0.15.2
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.3
wheel==0.43.0
win32-setctime==1.1.0
wrapt==1.16.0
zstandard==0.22.0
